requirements.txt
# fastapi==0.104.1
# uvicorn[standard]==0.24.0
# pydantic==2.5.0
# python-multipart==0.0.6
# httpx==0.25.2
# aiohttp==3.9.1
# asyncio==3.4.3
# python-dotenv==1.0.0
# pinecone-client==2.2.4
# openai==1.3.7
# google-generativeai==0.3.2
# anthropic==0.7.8
# transformers==4.36.0
# # torch==2.1.1
# numpy==1.24.3
# pandas==2.1.4
# # scikit-learn==1.3.2
# # langchain==0.0.350
# # langchain-community==0.0.10
# # langchain-openai==0.0.2
# # langchain-google-genai==0.0.6
# # langchain-anthropic==0.0.1
# structlog==23.2.0
# python-json-logger==2.0.7
# redis==5.0.1
# celery==5.3.4
# pytest==7.4.3
# pytest-asyncio==0.21.1
# black==23.11.0
# isort==5.12.0
# flake8==6.1.0
# mypy==1.7.1

# Core backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
aiohttp==3.9.1
python-dotenv==1.0.0

# Vector database & embeddings
pinecone-client==2.2.4
numpy==1.24.3
pandas==2.1.4

# LLM APIs
openai>=1.6.1,<2.0.0
google-generativeai==0.3.2
anthropic==0.7.8
transformers==4.36.0

# Logging
structlog==23.2.0
python-json-logger==2.0.7
redis==5.0.1

# Async tasks
celery==5.3.4

# Dev / testing / linting
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pydantic-settings
